+++ search/search.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"reflect"
+	"sort"
 	"strings"
 	"sync"
 
@@ -127,12 +128,13 @@ func getValue(s *string) string {
 }
 
 // mapToString converts a map[string]string to a string
-func mapToString(m map[string]string) string {
+func mapToString(m map[string]string, kvSep, listSep string) string {
 	var tags []string
 	for k, v := range m {
-		tags = append(tags, fmt.Sprintf("[%s=%s]", k, v))
+		tags = append(tags, fmt.Sprintf("%s%s%s", k, kvSep, v))
 	}
-	return strings.Join(tags, " ")
+	sort.Strings(tags)
+	return strings.Join(tags, listSep)
 }
 
 // ParseTags parses the tags to a map