+++ CHANGELOG.md
@@ -6,6 +6,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [v0.5.6] - 2022-11-25
+
+<!-- markdownlint-disable MD024 -->
+### Changed
+
+- Change function `mapToString` to accept 2 additional parameters: `kvSep` and `listSep` to allow customizing the separator between key and value and the separator between key-value pairs.
+- Rename config file option from `sort_by` to `sort` to be consistent with the command line option.
+- Rename config file option from `show_empty` to `show-empty` to be consistent with the command line option.
+- Rename config file option from `all_regions` to `all-regions` to be consistent with the other options.
+- Move config file option from `table_style` to `table.style`.
+- Tags are now sorted by default when using `--output table`.
+
+<!-- markdownlint-disable MD024 -->
+### Added
+
+- Add viper config for the `mapToString` parameters. You may add `separators.kv` and `separators.list` to the config file.
+
 ## [v0.5.5] - 2022-11-13
 
 <!-- markdownlint-disable MD024 -->